Sweeping changes to the PRISMS architecture. Too many new features, upgrades, and refactors to list. 3.0.0 is NOT backward-compatible, but well worth porting to.
Although many large changes have been made, a large portion of the reason that almost every file has been modified in this change set is due to either clean-up operations or the move from using org.dom4j.Element configurations throughout the PRISMS architecture to prisms.arch.PrismsConfig. The former is of course backward-compatible and the latter is easy for manual porting.