Fixed several bugs, including a serious parsing bug and a bug with generic type resolution.  Not sure generics is completely sound, but it's better.
Implemented the ablity to save and restore an evaluation environment.  Variables and history values may be saved as long as they implement java.io.Serializable and obey its contract.  Declared functions and imports may always be stored.  Use env.save(String) and env.load(String).  If an environment is saved as "default", it will be loaded on startup.